@@ -26,14 +26,14 @@
     "@ionic/storage": "2.1.3",
     "angular2-jwt": "^0.2.3",
     "ionic-angular": "3.9.2",
-    "ionicons": "3.0.0",
+    "ionicons": "4.0.0",
     "rxjs": "5.5.2",
     "sw-toolbox": "3.6.0",
     "zone.js": "0.8.18"
   },
   "devDependencies": {
     "@ionic/app-scripts": "^3.2.4",
-    "typescript": "2.4.2"
+    "typescript": "4.5.4"
   },
   "description": "An Ionic project"
 }
