@@ -1,9 +1,9 @@
-import {HttpClient} from "@angular/common/http";
-import {Injectable} from "@angular/core";
-import {Observable} from "rxjs/Rx";
-import {API_CONFIG} from "../../config/api.config";
-import {ClientDTO} from "../../models/client.dto";
-import {StorageService} from "../storage.service";
+import { HttpClient } from "@angular/common/http";
+import { Injectable } from "@angular/core";
+import { Observable } from "rxjs/Rx";
+import { API_CONFIG } from "../../config/api.config";
+import { ClientDTO } from "../../models/client.dto";
+import { StorageService } from "../storage.service";
 
 @Injectable()
 export class ClientService {
@@ -11,13 +11,13 @@ export class ClientService {
     constructor(private http: HttpClient, private storage: StorageService) {
     }
 
-    findByEmail(email: string): Observable<ClientDTO> {
-        return this.http.get<ClientDTO>(`${API_CONFIG.baseUrl}/clients/email?value=${email}`)
+    findByEmail(email: string) {
+        return this.http.get(`${API_CONFIG.baseUrl}/clients/email?value=${email}`);
     }
 
     getImageFromBucket(id: string): Observable<any> {
         let url = `${API_CONFIG.bucketBaseUrl}/cp${id}.jpg`
-        return this.http.get(url, {responseType: "blob"})
+        return this.http.get(url, { responseType: "blob" })
     }
 
     insert(obj: ClientDTO) {
